[broken] split graph.py into 3 files
