+++ topos/base/__init__.py
@@ -1,4 +1,6 @@
 from .graph         import Graph
+from .complex       import Complex
+from .nerve         import Nerve
 # old
 #from .set           import Set, Iterable, Map, Mapping
 #from .fiber         import Fiber
