+++ topos/base/graph.py
@@ -1,5 +1,8 @@
-from topos.core import sparse, Shape, SymGroup
+from topos.core import sparse, Shape
 from topos.io   import alignString, parseTensor
+
+import topos.base.nerve    
+import topos.base.complex  
 
 import torch
 from torch import stack, cat, arange
@@ -107,7 +110,7 @@ class Graph :
             Nd = sparse.matrix([Ntot] * (deg + 1), stack(ijk))
             N += [Nd.coalesce()]
             deg += 1
-        return Nerve(*(Nd.indices().T for Nd in N), sort=False)
+        return topos.base.nerve.Nerve(*(Nd.indices().T for Nd in N), sort=False)
 
     def arrows (self): 
         """ 1-Chains of the hypergraph. """
@@ -124,7 +127,7 @@ class Graph :
             # row indices
             i_ = I[n].index_select(0, E[n].index(Gn)).to_dense()
             # loop over subfaces
-            F  = simplices(Gn)
+            F  = topos.base.complex.Complex.simplices(Gn)
             for k, Fk in enumerate(F[:-1]):
                 # valid column indices
                 nz = (A[k].index_select(0, E[k].index(Fk))
@@ -159,148 +162,3 @@ class Graph :
             s += alignString(degree(d), self[d]) \
                 + (',\n\n' if d < self.dim else '\n\n')
         return s + '}'
-
-
-class Complex (Graph):
-    """
-    Simplicial complexes. 
-
-    A simplicial complex is a hypergraph containing all the subsets of its cells. 
-    The i-th face map consists of removing the i-th vertex from a degree-n cell,
-    for 0 <= i <= n.
-    Alternated sums of face maps induce a dual pair of differential and codifferential 
-    (or boundary) operators, `d` and `delta = d*`. 
-
-    References:
-    -----------
-    - Peltre, 2020:
-        Message-Passing Algorithms and Homology, Chapter II
-        https://arxiv.org/abs/2009.11631
-    """
-
-    def diff(self, d):
-        """ Differential d: K[d] -> K[d + 1]. """
-        src, tgt = self[d], self[d + 1]
-        N, P = tgt.shape[0], src.shape[0]
-        i = self.index(tgt) - self.index(tgt[0])
-        out = sparse.matrix([N, P], [])
-        for k in range(d + 2):
-            j0  = self.index(src[0])
-            j   = self.index(face(k, self[d + 1])) - j0
-            val = (-1.) ** k
-            out += sparse.matrix([N, P], stack([i, j]), val, t=0)
-        return out
-    
-    @staticmethod
-    def face(i, f):
-        """ Face map acting on (batched) cells [j0, ..., jn]. """
-        n = f.shape[-1]
-        return f.index_select(f.dim() - 1, cat([arange(i), arange(i+1, n)]))
-    
-    @classmethod
-    def simplicial(cls, faces):
-        """ Simplicial closure. 
-            
-            Returns the complex containing every subface of the input faces.
-        """
-        faces = parseTensor(faces)
-        K = [[(0, faces)]]
-        def nextf (nfaces):
-            faces = []
-            for i, f in nfaces:
-                faces += [(j, cls.face(j, f)) for j in range(i, f.shape[-1])]
-            return faces
-        for codim in range(faces.shape[-1] - 1):
-            K += [nextf(K[-1])]
-        K_ = [cat([f for i, f in Kn[::-1]]) for Kn in K[::-1]]
-        return cls(*K_)
-            
-    def __repr__(self):
-        return f'{self.dim} Complex {self}'
-
-
-class Nerve (Complex):
-    """
-    Categorical nerves. 
-
-    The nerve of a hypergraph is a simplicial complex. 
-    The partial order structure of the hypergraph also
-    induces a pair of combinatorial automorphisms, 
-    extending the Zeta transform and Möbius inversion formula
-    to the whole complex. 
-
-    References:
-    -----------
-    - Rota, 1964: 
-        On the Foundations of Combinatorial Theory - I. Theory of Möbius Transforms
-    - Peltre, 2020: 
-        Message-Passing Algorithms and Homology, Chapter III, 
-        https://arxiv.org/abs/2009.11631
-    """
-    
-    
-    def zeta (self, d):
-        """ Degree-d zeta transform. 
-
-            See Nerve.zetas(d) to keep graded components until d.  
-        """
-        return self.zetas(d)[d]
-
-    def zetas (self, d):
-        """ List of zeta transforms in degrees <= d. """ 
-
-        zt, N = [], [Nd.shape[0] for Nd in self.grades]
-
-        # strict inclusions
-        A  = self.adj[1]
-        # weak inclusions
-        zt0 = ( sparse.matrix([N[0], N[0]], A.indices(), t=False) 
-              + sparse.eye(N[0])).coalesce()
-        
-        zt += [zt0]
-        
-        def next_diagrams (chain_i, chain_j):
-            """ Extend diagrams to the right:
-
-                    i0 > i1 > ... > id
-                    >=   >=         >=
-                    j0 > j1 > ... > jd
-            """
-            order = sparse.reshape([-1], zt0).coalesce()
-            # Extend chain_i
-            below_i = sparse.index_select(A, chain_i[-1])
-            next_i  = below_i.indices()[1]
-            deg_i   = sparse.sum_dense(below_i, [1]).long()
-            chain_j = chain_j.repeat_interleave(deg_i, 1)
-            chain_i = chain_i.repeat_interleave(deg_i, 1)
-            chain_i = torch.cat([chain_i, next_i[None,:]])
-            # Extend chain_j
-            below_j = sparse.index_select(A, chain_j[-1])
-            next_j  = below_j.indices()[1]
-            deg_j   = sparse.sum_dense(below_j, [1]).long()
-            chain_i = chain_i.repeat_interleave(deg_j, 1)
-            chain_j = chain_j.repeat_interleave(deg_j, 1)
-            chain_j = torch.cat([chain_j, next_j[None,:]])
-            # Check that i[d] >= j[d] and i[d] !>= j[d-1]
-            idx1  = N[0] * chain_i[-1] + chain_j[-1]
-            idx2  = N[0] * chain_i[-1] + chain_j[-2]
-            mask1 =   sparse.index_mask(order, idx1)
-            mask2 = ~ sparse.index_mask(order, idx2)
-            nz = (mask1 * mask2).nonzero().flatten()
-            # Return diagrams
-            return chain_i[:, nz], chain_j[:, nz]
-
-        acc = [zt0.indices().unsqueeze(1)]
-
-        for k in range(d):
-            acc += [torch.stack(next_diagrams(*acc[-1]))]
-        
-        for chains, n, begin in zip(acc[1:], N[1:d+1], self.sizes[:d]):
-            ij = torch.stack([self.index(chains[0].T) - begin, 
-                              self.index(chains[1].T) - begin])
-            zt += [sparse.tensor([n, n], ij, t=False)]
-
-        return zt
-
-    def __repr__(self):
-        return f'{self.dim} Nerve {self}'
